+++ src/HTTP.ts
@@ -7,6 +7,7 @@ import * as Header from "./HTTP/Header";
 import * as JSONParser from "./HTTP/JSONParser";
 import * as JSONLDParser from "./HTTP/JSONLDParser";
 import Method from "./HTTP/Method";
+import * as Parser from "./HTTP/Parser";
 import * as Request from "./HTTP/Request";
 import * as Response from "./HTTP/Response";
 import StatusCode from "./HTTP/StatusCode";
@@ -18,6 +19,7 @@ export {
 	JSONParser,
 	JSONLDParser,
 	Method,
+	Parser,
 	Request,
 	Response,
 	StatusCode,