+++ src/NS/C.ts
@@ -3,6 +3,11 @@ export let namespace:string = "https://carbonldp.com/ns/v1/platform#";
 export class Class {
 	static get AccessPoint():string { return namespace + "AccessPoint"; }
 	static get API():string { return namespace + "API"; }
+	static get NonReadableMembershipResourceTriples():string { return namespace + "NonReadableMembershipResourceTriples"; }
+	static get PreferContainmentResources():string { return namespace + "PreferContainmentResources"; }
+	static get PreferContainmentTriples():string { return namespace + "PreferContainmentTriples"; }
+	static get PreferMembershipResources():string { return namespace + "PreferMembershipResources"; }
+	static get PreferMembershipTriples():string { return namespace + "PreferMembershipTriples"; }
 	static get VolatileResource():string { return namespace + "VolatileResource"; }
 }